datitos

+++ funciones/CF_it.py
@@ -9,7 +9,7 @@ def cf_it():
         a = 1
         for linea in range(0, len(a_data), 20):
             aux2 = list(filter(lambda x: x[0] == a_data[linea][0], a_data))
-            aux2 = list(map(lambda x: [int(x[1]), int(x[2])], aux2))
+            aux2 = list(map(lambda x: [int(x[1]), float(x[2])], aux2))
             aux.append([False] + aux2)
 
         a_data = [False] + aux
